Huge refactoring, I'mma be real shoutout Copilot men jag har kollat igenom koden och manuellt ändrat smågrejer
