+++ src/app/components/admin/index.ts
@@ -1,0 +1,14 @@
+// Export all admin components for easier importing
+export { default as CreateEventForm } from './CreateEventForm';
+export { default as ManageEvents } from './ManageEvents';
+export { default as CreatePostForm } from './CreatePostForm';
+export { default as UpdateDisplayNameForm } from './UpdateDisplayNameForm';
+export { default as UpdateProfilePictureForm } from './UpdateProfilePictureForm';
+export { default as ManageAdmins } from './ManageAdmins';
+export { default as UploadBlandareForm } from './UploadBlandareForm';
+export { default as ManageBlandare } from './ManageBlandare';
+
+// Export sub-components (optional - for direct use if needed)
+export { default as ImageCropper } from './ImageCropper';
+export { default as EditEventModal } from './EditEventModal';
+export { default as EventsList } from './EventsList';
